refactor: product 세부분리
